Remove file-type from package.json
