+++ package.json
@@ -65,7 +65,6 @@
     "catbox-s3": "^4.0.0",
     "cheerio": "^1.0.0-rc.3",
     "config": "^1.30.0",
-    "file-type": "^12.0.1",
     "good": "^8.1.1",
     "good-console": "^7.1.0",
     "good-squeeze": "^5.0.2",
